+++ Revit.Busy/Properties/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+﻿using System;
+using System.Windows.Markup;
+[assembly: XmlnsPrefix("http://schemas.revit.busy.com/2024/xaml/presentation", "revit.busy")]
+[assembly: XmlnsDefinition("http://schemas.revit.busy.com/2024/xaml/presentation", "Revit.Busy")]