+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.2.2] / 2023-08-14
+### Features
+- Add `XmlnsPrefix` and `XmlnsDefinition` to `Revit.Busy` namespace.
 ### Updated
 - Update references to work with net code.
 - Clear main project and move files to `Example`